// my-vite-app/src/types/types.ts
export interface AccountDetails {
  accountno: string;
  status: string;
  telephoneno: string;
}